دعم مبدئي لمكتبة الرياضيات

+++ src/Models/AlifBuild.js
@@ -43,9 +43,12 @@ export async function بناء_الف(fileName) {
 
         const مجلد_الويب = path.join(process.cwd(), "بناء", "ويب");
         fs.mkdirSync(مجلد_الويب, { recursive: true });
-        fs.cpSync("المرفقات", path.join(مجلد_الويب, "المرفقات"), {
-            recursive: true,
-        });
+
+        if (fs.existsSync("المرفقات")) {
+            fs.cpSync("المرفقات", path.join(مجلد_الويب, "المرفقات"), {
+                recursive: true,
+            });
+        }
 
         fs.writeFileSync(
             path.join(مجلد_الويب, `${اسم_المف}.alif.js`),
@@ -75,7 +78,7 @@ export async function بناء_الف(fileName) {
             "utf8"
         );
 
-        console.log("✅ تم بناء المشروع وتحسينه داخل بناء/ويب/");
+        console.log("تم بناء المشروع وتحسينه داخل بناء/ويب/");
     } catch (e) {
         console.error("خطأ أثناء التوليد:", e.message);
     }
